Reviewed-on: https://egit.irs.uni-stuttgart.de/rust/spacepackets/pulls/5

+++ src/time.rs
@@ -134,7 +134,7 @@ impl CdsShortTimeProvider {
         };
         let unix_days_seconds =
             ccsds_to_unix_days(ccsds_days as i32) as i64 * SECONDS_PER_DAY as i64;
-        provider.setup(unix_days_seconds as i64, ms_of_day.into())
+        provider.setup(unix_days_seconds, ms_of_day.into())
     }
 
     #[cfg(feature = "std")]
@@ -223,7 +223,7 @@ impl CcsdsTimeProvider for CdsShortTimeProvider {
     }
 
     fn date_time(&self) -> Option<DateTime<Utc>> {
-        self.calc_date_time((self.ms_of_day % 1000) as u32)
+        self.calc_date_time(self.ms_of_day % 1000)
     }
 }
